Moved requests to internal package GetUpdates now works via POST
