+++ send/internal/requests/get_updates.go
@@ -1,0 +1,7 @@
+package send_requests
+
+type GetUpdatesType struct {
+	Offset int64 `json:"offset,omitempty"`
+	Limit int64 `json:"limit,omitempty"`
+	Timeout int64 `json:"timeout,omitempty"`
+}
